+++ frontend/src/css/startEvent.css
@@ -1,0 +1,6 @@
+.startpark {
+	max-width: 100px;
+	height: 10px;
+	padding: 1rem;
+	color: blue;
+}
